+++ src/main/java/com/hackathonteam1/refreshrator/service/RecipeServiceImpl.java
@@ -1,6 +1,8 @@
 package com.hackathonteam1.refreshrator.service;
 
+import com.hackathonteam1.refreshrator.dto.request.recipe.DeleteIngredientRecipesDto;
 import com.hackathonteam1.refreshrator.dto.request.recipe.ModifyRecipeDto;
+import com.hackathonteam1.refreshrator.dto.request.recipe.RegisterIngredientRecipesDto;
 import com.hackathonteam1.refreshrator.dto.request.recipe.RegisterRecipeDto;
 import com.hackathonteam1.refreshrator.dto.response.recipe.DetailRecipeDto;
 import com.hackathonteam1.refreshrator.entity.*;
@@ -41,9 +43,9 @@ import java.util.stream.Collectors;
 @Transactional
 public class RecipeServiceImpl implements RecipeService{
     private final ImageService imageService;
-    private final IngredientRecipeServiceImpl ingredientRecipeServiceImpl;
     private final UserService userService;
     private final FridgeService fridgeService;
+    private final IngredientService ingredientService;
 
     private final RecipeRepository recipeRepository;
     private final IngredientRecipeRepository ingredientRecipeRepository;
@@ -85,7 +87,7 @@ public class RecipeServiceImpl implements RecipeService{
         List<UUID> ingredientIds = registerRecipeDto.getIngredientIds();
 
         //동일한 재료를 요청할 경우 예외처리
-        ingredientRecipeServiceImpl.checkDuplicatedIngredient(ingredientIds);
+        checkDuplicatedIngredient(ingredientIds);
 
         Recipe recipe = Recipe.builder()
                 .name(registerRecipeDto.getName())
@@ -96,7 +98,7 @@ public class RecipeServiceImpl implements RecipeService{
 
         recipeRepository.save(recipe);
 
-        ingredientRecipeServiceImpl.registerIngredients(ingredientIds, recipe);
+        registerIngredients(ingredientIds, recipe);
     }
 
     //상세조회
@@ -215,6 +217,63 @@ public class RecipeServiceImpl implements RecipeService{
         return recipeListDto;
     }
 
+    //레시피 재료 등록
+    @Transactional
+    @CacheEvict(value = "recipeDetailCache", key = "#recipeId", cacheManager = "redisCacheManager")
+    public void registerIngredientRecipe(User user, UUID recipeId, RegisterIngredientRecipesDto registerIngredientRecipesDto) {
+        Recipe recipe = findRecipeById(recipeId);
+
+        if(!userService.isAuthorized(recipe.getUser(), user)){ //레피시 작성자 여부 확인
+            throw new ForbiddenException(ErrorCode.RECIPE_FORBIDDEN);
+        }
+
+        //기존에 레시피에 존재하던 재료 리스트, contains 여부 속도 향상을 위해 Set을 사용
+        Set<Ingredient> existingIngredients = new HashSet<>(findAllIngredientByIngredientRecipes(findAllIngredientRecipeByRecipe(recipe)));
+
+        List<UUID> requestedIngredientIds = registerIngredientRecipesDto.getIngredientIds();
+        //요청된 재료에 중복이 있는지 확인
+        checkDuplicatedIngredient(requestedIngredientIds);
+
+        List<Ingredient> newIngredients = requestedIngredientIds.stream().map(ingredientService::findIngredientById).toList();
+
+        //기존에 레시피에 존재하던 재료인지 확인 후 추가
+        newIngredients.forEach(newIngredient->{
+            if(existingIngredients.contains(newIngredient)){
+                throw new ConflictException(ErrorCode.RECIPE_INGREDIENT_CONFLICT);
+            }
+            registerRecipeIngredient(newIngredient, recipe);
+        });
+    }
+
+    //레시피 재료 삭제
+    @CacheEvict(value = "recipeDetailCache", key = "#recipeId", cacheManager = "redisCacheManager")
+    public void deleteIngredientRecipe(User user, UUID recipeId, DeleteIngredientRecipesDto deleteIngredientRecipesDto) {
+        Recipe recipe = findRecipeById(recipeId);
+
+        if(!userService.isAuthorized(recipe.getUser(), user)){ //레피시 작성자 여부 확인
+            throw new ForbiddenException(ErrorCode.RECIPE_FORBIDDEN);
+        }
+
+        //기존 레시피의 재료들
+        List<IngredientRecipe> existingIngredientRecipes = findAllIngredientRecipeByRecipe(recipe);
+        //기존 레시피의 재료들의 Id를 파싱해서 가져옴
+        Set<UUID> existingIngredientRecipeIds = existingIngredientRecipes.stream().map(i->i.getId()).collect(Collectors.toSet());
+
+        List<UUID>requestedIngredientIds = deleteIngredientRecipesDto.getIngredientRecipeIds();
+
+        //요청된 재료 중복 확인
+        checkDuplicatedIngredient(requestedIngredientIds);
+
+        requestedIngredientIds.forEach(i -> {
+            if(!existingIngredientRecipeIds.contains(i)){
+                throw new NotFoundException(ErrorCode.INGREDIENT_RECIPE_NOT_FOUND);
+            }
+            ingredientRecipeRepository.deleteById(i);
+        });
+
+    }
+
+
     // 레시피에 좋아요 추가
     @Override
     @Caching(
@@ -294,6 +353,34 @@ public class RecipeServiceImpl implements RecipeService{
             return Sort.by(Sort.Order.desc("likeCount"));
         }
         throw new BadRequestException(ErrorCode.SORT_TYPE_ERROR);
+    }
+
+    //레시피에 재료들 등록
+    private void registerIngredients(List<UUID> ingredientIds, Recipe recipe){
+        ingredientIds.forEach(ingredientId->
+                registerRecipeIngredient(ingredientService.findIngredientById(ingredientId), recipe));
+    }
+
+    //요청에 중복된 재료가 있는지 확인.
+    private void checkDuplicatedIngredient(List<UUID> ingredientIds){
+        Set<UUID> ingredientIdSet = new HashSet<>(ingredientIds);
+        if(ingredientIdSet.size() != ingredientIds.size()){
+            throw new BadRequestException(ErrorCode.DUPLICATED_RECIPE_INGREDIENT);
+        }
+    }
+
+    //RecipeIngredient를 등록하는 메서드
+    private void registerRecipeIngredient(Ingredient ingredient, Recipe recipe){
+        IngredientRecipe ingredientRecipe = IngredientRecipe.builder()
+                .recipe(recipe)
+                .ingredient(ingredient)
+                .build();
+        ingredientRecipeRepository.save(ingredientRecipe);
+    }
+
+    //IngredientRecipe 리스트로 Ingredient 리스트 생성하는 메서드
+    private List<Ingredient> findAllIngredientByIngredientRecipes(List<IngredientRecipe> ingredientRecipes){
+        return ingredientRecipes.stream().map(i->i.getIngredient()).collect(Collectors.toList());
     }
 
 }
